Car Park application
the developed mobile application is used to provide the functionality to the user to book the parking slot at the desired location for the preferred time slot. This mobile application enables the user to set the directions to the selected parking lot location using the Google map location services. In this application we have provided two types of user logins which are, customer and parking inspector.